feat: move account and character to common types
