+++ electron/common/account/types.ts
@@ -1,0 +1,10 @@
+export interface Account {
+  accountName: string;
+  accountPassword: string;
+}
+
+export interface Character {
+  accountName: string;
+  characterName: string;
+  gameCode: string;
+}
